警告レベル3をクリア

+++ Src/Engine/DirectXBase.cpp
@@ -151,7 +151,7 @@ void DirectXBase::Init() {
 		D3D12_HEAP_FLAG_NONE,
 		&depthResDesc,
 		D3D12_RESOURCE_STATE_DEPTH_WRITE,
-		&CD3DX12_CLEAR_VALUE(DXGI_FORMAT_D32_FLOAT, 1, 0.0f),
+		&CD3DX12_CLEAR_VALUE(DXGI_FORMAT_D32_FLOAT, static_cast<FLOAT>(1), static_cast<UINT8>(0)),
 		IID_PPV_ARGS(&depthBuffer)
 	);
